feat: Add open sockets stats
